refactor advisory policy review queue router
