+++ src/app/routers/advisory_policy_evaluations.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Header, Path, Query
+from fastapi import APIRouter, Header, Path
 
 from app.contracts.advisory_policy import (
     AdvisoryPolicyBodyRequest,
@@ -39,33 +39,5 @@ async def create_policy_evaluation(
         proposal_version_id=proposal_version_id,
         body=request.body,
         idempotency_key=idempotency_key,
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.get(
-    "/advisory-policy-evaluations/review-queue",
-    response_model=AdvisoryPolicyEnvelopeResponse,
-    summary="List Advisory Policy Review Queue",
-    description=(
-        "Returns policy evaluation review-queue items from lotus-advise for advisor, "
-        "compliance, investment desk, operations, and supervisory workflows."
-    ),
-)
-async def get_policy_review_queue(
-    evaluation_status: str | None = Query(
-        default=None,
-        description="Optional policy evaluation status filter owned by lotus-advise.",
-        examples=["PENDING_REVIEW"],
-    ),
-    portfolio_id: str | None = Query(
-        default=None,
-        description="Optional portfolio identifier filter owned by lotus-advise.",
-        examples=["PB_SG_GLOBAL_BAL_001"],
-    ),
-) -> AdvisoryPolicyEnvelopeResponse:
-    return await advisory_policy_service().get_policy_review_queue(
-        evaluation_status=evaluation_status,
-        portfolio_id=portfolio_id,
         correlation_id=correlation_id_var.get(),
     )
